\name{wiki_graph}
\docType{data}
\alias{wiki_graph}
\title{Graph example from Wikipedia}
\description{
  This data set is an example graph from the Wikipedia page about Dijkstra's
  algorithm. It contains three variables v1, v2 and w.
}
\usage{wiki_graph}
\format{A data.frame containing three numeric columns.}
\source{\url{https://en.wikipedia.org/wiki/Dijkstra\%27s algorithm}}
\references{
  \url{https://en.wikipedia.org/wiki/Dijkstra\%27s algorithm}
}
\keyword{datasets}